microservices/pipeline/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.4.26
cffi==1.17.1
dnspython==2.7.0
laspy==2.5.4
laszip==0.2.3
lazrs==0.6.2
minio==7.2.15
numpy==2.2.5
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pymongo==4.12.1
pyproj==3.7.1
python-dotenv==1.1.0
scipy==1.15.3
typing_extensions==4.13.2
urllib3==2.4.0
